requirements.txt
Flask==3.0.2
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
kafka-python==2.0.2
Werkzeug==3.0.1
SQLAlchemy==2.0.28
python-socketio==5.11.1
eventlet==0.35.2
gunicorn==21.2.0
pytest==8.0.2
pytest-cov==4.1.0
black==24.2.0
flake8==7.0.0
mypy==1.8.0 